requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.2
xgboost==2.0.3
scikit-learn==1.3.2
plotly==5.18.0
ccxt==4.2.25
huggingface-hub==0.20.3
pyarrow==15.0.0
ta==0.11.0
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
joblib==1.3.2
